Improve commenting + general refactoring
