+++ osu.Framework/Input/Handlers/Mouse/OpenTKMouseHandler.cs
@@ -81,4 +81,3 @@ namespace osu.Framework.Input.Handlers.Mouse
         }
     }
 }
-
