+++ osu.Framework/Input/IMouseState.cs
@@ -17,10 +17,12 @@ namespace osu.Framework.Input
         Vector2 Position { get; set; }
 
         /// <summary>
-        /// Whether if this mouse state contains a valid <see cref="Position"/>.
+        /// Whether this <see cref="IMouseState"/> contains a valid <see cref="Position"/>.
         /// If this is false, the value of <see cref="Position"/> should not be used.
-        /// An example of a state with an invalid position is at before the first mouse input is handled.
         /// </summary>
+        /// <remarks>
+        /// For example, if the first mouse input has not been handled, the <see cref="Position"/> will be invalid.
+        /// </remarks>
         bool IsPositionValid { get; set; }
 
         Vector2 LastPosition { get; set; }
