+++ osu.Framework/Input/MouseScrollRelativeInput.cs
@@ -7,18 +7,20 @@ namespace osu.Framework.Input
 {
     /// <summary>
     /// Denotes a relative change of mouse scroll.
-    /// Normal devices such as mouses have relative scrolling capability.
+    /// Pointing devices such as mice provide relative scroll input.
     /// </summary>
     public class MouseScrollRelativeInput : IInput
     {
         /// <summary>
-        /// The delta value which will be added to the current mouse scroll.
+        /// The change in scroll. This is added to the current scroll.
         /// </summary>
         public Vector2 Delta;
+
         /// <summary>
-        /// Whether it is a precise scrolling.
+        /// Whether the change occurred as the result of a precise scroll.
         /// </summary>
         public bool IsPrecise;
+
         public void Apply(InputState state, IInputStateChangeHandler handler)
         {
             var mouse = state.Mouse;
